Split tests for binary ops from test_math.py
