+++ tests/chainerx_tests/unit_tests/routines_tests/test_math.py
@@ -1,5 +1,3 @@
-import unittest
-
 import chainer
 import numpy
 import pytest
@@ -9,178 +7,8 @@ import chainerx.testing
 
 from chainerx_tests import array_utils
 from chainerx_tests import dtype_utils
+from chainerx_tests import math_utils
 from chainerx_tests import op_utils
-
-
-class IgnoreNumpyFloatingPointError(object):
-
-    def __enter__(self):
-        self.old_settings = numpy.seterr(all='ignore')
-
-    def __exit__(self, *args):
-        numpy.seterr(**self.old_settings)
-
-
-class UnaryMathTestBase(object):
-
-    input = None
-
-    def setup(self):
-        in_dtype, = self.in_dtypes
-        in_kind = numpy.dtype(in_dtype).kind
-
-        if numpy.dtype(in_dtype).kind != 'f':
-            self.skip_backward_test = True
-            self.skip_double_backward_test = True
-
-        if in_dtype == 'float16':
-            self.check_forward_options.update({'rtol': 1e-3, 'atol': 1e-3})
-            self.check_backward_options.update({'rtol': 3e-3, 'atol': 3e-3})
-            self.check_double_backward_options.update(
-                {'rtol': 1e-2, 'atol': 1e-2})
-
-        input = self.input
-        if (in_kind == 'u'
-                and isinstance(input, (int, float))
-                and input < 0):
-            raise unittest.SkipTest(
-                'Combination of uint dtype and negative input cannot be '
-                'tested')
-
-    def generate_inputs(self):
-        in_dtype, = self.in_dtypes
-        if isinstance(self.input, numpy.ndarray):
-            return self.input.astype(in_dtype),
-        if self.input == 'random':
-            return array_utils.uniform(self.shape, in_dtype),
-        if isinstance(self.input, (bool, int, float)):
-            return numpy.full(self.shape, self.input, dtype=in_dtype),
-        assert False
-
-    def forward_xp(self, inputs, xp):
-        a, = inputs
-        # This cast was introduced in order to avoid decreasing precision.
-        # ex.) numpy.sqrt(x) becomes a float16 array where x is an int8 array.
-        a = dtype_utils.cast_if_numpy_array(xp, a, self.out_dtype)
-        with IgnoreNumpyFloatingPointError():
-            y = self.func(xp, a)
-        y = dtype_utils.cast_if_numpy_array(xp, y, self.out_dtype)
-        return y,
-
-
-class BinaryMathTestBase(object):
-
-    def setup(self):
-        in_dtype1, in_dtype2 = self.in_dtypes
-
-        kind1 = numpy.dtype(in_dtype1).kind
-        kind2 = numpy.dtype(in_dtype2).kind
-        if kind1 != 'f' or kind2 != 'f':
-            self.skip_backward_test = True
-            self.skip_double_backward_test = True
-
-        if in_dtype1 == 'float16' or in_dtype2 == 'float16':
-            self.check_forward_options.update({'rtol': 1e-3, 'atol': 1e-3})
-            self.check_backward_options.update({'rtol': 1e-3, 'atol': 1e-3})
-            self.check_double_backward_options.update(
-                {'rtol': 1e-3, 'atol': 1e-3})
-
-    def generate_inputs(self):
-        in_dtype1, in_dtype2 = self.in_dtypes
-        in_shape1, in_shape2 = self.in_shapes
-        if self.input_lhs == 'random':
-            a = array_utils.uniform(in_shape1, in_dtype1)
-        elif isinstance(self.input_lhs, (bool, int, float)):
-            a = numpy.full(in_shape1, self.input_lhs, dtype=in_dtype1)
-        else:
-            assert False
-        if self.input_rhs == 'random':
-            b = array_utils.uniform(in_shape2, in_dtype2)
-        elif isinstance(self.input_rhs, (bool, int, float)):
-            b = numpy.full(in_shape2, self.input_rhs, dtype=in_dtype2)
-        else:
-            assert False
-        return a, b
-
-    def forward_xp(self, inputs, xp):
-        a, b = inputs
-        # This cast was introduced in order to avoid decreasing precision.
-        # ex.) x / y becomes a float16 array where x and y are an int8 arrays.
-        a = dtype_utils.cast_if_numpy_array(xp, a, self.out_dtype)
-        b = dtype_utils.cast_if_numpy_array(xp, b, self.out_dtype)
-        with IgnoreNumpyFloatingPointError():
-            y = self.func(xp, a, b)
-        y = dtype_utils.cast_if_numpy_array(xp, y, self.out_dtype)
-        return y,
-
-
-class InplaceUnaryMathTestBase(UnaryMathTestBase):
-
-    skip_backward_test = True
-    skip_double_backward_test = True
-
-    def forward_xp(self, inputs, xp):
-        a, = inputs
-        if xp is chainerx:
-            a_ = a.as_grad_stopped().copy()
-        else:
-            a_ = a.copy()
-        with IgnoreNumpyFloatingPointError():
-            ret = self.func(xp, a_)
-        assert ret is None  # func should not return anything
-        return a_,
-
-
-class InplaceBinaryMathTestBase(BinaryMathTestBase):
-
-    skip_backward_test = True
-    skip_double_backward_test = True
-
-    def forward_xp(self, inputs, xp):
-        a, b = inputs
-        b = dtype_utils.cast_if_numpy_array(xp, b, a.dtype)
-        if xp is chainerx:
-            a_ = a.as_grad_stopped().copy()
-            b_ = b.as_grad_stopped()
-        else:
-            a_ = a.copy()
-            b_ = b
-        with IgnoreNumpyFloatingPointError():
-            ret = self.func(xp, a_, b_)
-        assert ret is None  # func should not return anything
-        return a_,
-
-
-def _convert_numpy_scalar(scalar, dtype):
-    # Implicit casting in NumPy's multiply depends on the 'casting' argument,
-    # which is not yet supported (ChainerX always casts).
-    # Therefore, we explicitly cast the scalar to the dtype of the ndarray
-    # before the multiplication for NumPy.
-    return numpy.dtype(dtype).type(scalar)
-
-
-class MathScalarTestBase(UnaryMathTestBase):
-
-    def func(self, xp, a):
-        scalar = self.scalar_type(self.scalar_value)
-        return self.func_scalar(xp, a, scalar)
-
-
-class InplaceMathScalarTestBase(InplaceUnaryMathTestBase):
-
-    def func(self, xp, a):
-        scalar = self.scalar_type(self.scalar_value)
-        if xp is numpy:
-            # This cast is to avoid TypeError in the following case
-            #     a: uint8 0-dim numpy.ndarray
-            #     scalar: int
-            in_dtype, = self.in_dtypes
-            scalar = _convert_numpy_scalar(scalar, in_dtype)
-        return self.func_scalar(xp, a, scalar)
-
-
-def _make_same_in_out_dtypes(number_of_in_params, dtypes):
-    return [((dtype,) * number_of_in_params, dtype) for dtype in dtypes]
 
 
 _in_out_dtypes_arithmetic_invalid = [
@@ -283,59 +111,14 @@ _in_out_dtypes_inplace_float_arithmetic_scalar = (
     _in_out_dtypes_float_array_float_scalar)
 
 
-def _permutate_shapes(shapes_list):
-    # Permutates input shapes
-    permutated_shapes_list = []
-    for in_shape1, in_shape2 in shapes_list:
-        permutated_shapes_list.append((in_shape1, in_shape2))
-        permutated_shapes_list.append((in_shape2, in_shape1))
-    return list(set(permutated_shapes_list))
-
-
-_shapes_combination_inplace_binary = [
-    # Same shapes
-    ((1,), (1,)),
-    ((3, 4), (3, 4)),
-    # Broadcast
-    ((10,), (1,)),
-    ((3, 4), (3, 1)),
-    ((3, 4), (1, 4)),
-    ((3, 4), (4,)),
-    ((3, 4), (1, 1)),
-    ((3, 4), (1,)),
-    ((2, 3, 4), (1, 1, 1)),
-    # 0-dim shape
-    ((), ()),
-    ((1,), ()),
-    ((3,), ()),
-    ((2, 3), ()),
-    # 0-size shape
-    ((0,), (0,)),
-    ((0,), (1,)),
-    ((0,), ()),
-    ((2, 0, 3), (2, 0, 3)),
-    # TODO(imanishi): Fix strides
-    # ((2, 0, 3), (0, 1)),
-]
-
-
-_shapes_combination_binary = _permutate_shapes([
-    # Broadcast
-    ((3, 1), (1, 4)),
-    ((2, 1, 4), (3, 1)),
-    # 0-size shape
-    # TODO(imanishi): Fix strides
-    # ((0, 1), (0, 1, 0)),
-]) + _permutate_shapes(_shapes_combination_inplace_binary)
-
-
 @op_utils.op_test(['native:0', 'cuda:0'])
 @chainer.testing.parameterize(*(
     # Special shapes
     chainer.testing.product({
         'shape': [(), (0,), (1,), (2, 0, 3), (1, 1, 1), (2, 3)],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(1, chainerx.testing.numeric_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                1, chainerx.testing.numeric_dtypes)),
         'input': ['random'],
         'is_module': [False],
     })
@@ -343,7 +126,8 @@ _shapes_combination_binary = _permutate_shapes([
     + chainer.testing.product({
         'shape': [(2, 3)],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(1, chainerx.testing.numeric_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                1, chainerx.testing.numeric_dtypes)),
         'input': ['random'],
         'is_module': [True, False],
     })
@@ -351,14 +135,15 @@ _shapes_combination_binary = _permutate_shapes([
     + chainer.testing.product({
         'shape': [(2, 3)],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(1, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                1, chainerx.testing.float_dtypes)),
         'input': [float('inf'), -float('inf'), float('nan')],
         'is_module': [False],
         'skip_backward_test': [True],
         'skip_double_backward_test': [True],
     })
 ))
-class TestNegative(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestNegative(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         if self.is_module:
@@ -382,9 +167,10 @@ def test_negative_invalid_bool(xp, device, is_module):
 @chainer.testing.parameterize(*(
     # Special shapes
     chainer.testing.product({
-        'in_shapes': _shapes_combination_binary,
+        'in_shapes': math_utils.shapes_combination_binary,
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.numeric_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.numeric_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
         'is_module': [False],
@@ -401,7 +187,8 @@ def test_negative_invalid_bool(xp, device, is_module):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.numeric_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.numeric_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
         'is_module': [True, False],
@@ -410,7 +197,8 @@ def test_negative_invalid_bool(xp, device, is_module):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': ['random', float('inf'), -float('inf'), float('nan')],
         'input_rhs': ['random', float('inf'), -float('inf'), float('nan')],
         'is_module': [False],
@@ -418,7 +206,7 @@ def test_negative_invalid_bool(xp, device, is_module):
         'skip_double_backward_test': [True],
     })
 ))
-class TestAdd(BinaryMathTestBase, op_utils.NumpyOpTest):
+class TestAdd(math_utils.BinaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a, b):
         if self.is_module:
@@ -445,9 +233,10 @@ def test_add_invalid_dtypes(device, dtypes, is_module):
 @chainer.testing.parameterize(*(
     # Special shapes
     chainer.testing.product({
-        'in_shapes': _shapes_combination_inplace_binary,
+        'in_shapes': math_utils.shapes_combination_inplace_binary,
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.numeric_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.numeric_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
     })
@@ -462,14 +251,15 @@ def test_add_invalid_dtypes(device, dtypes, is_module):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': ['random', float('inf'), -float('inf'), float('nan')],
         'input_rhs': ['random', float('inf'), -float('inf'), float('nan')],
         'skip_backward_test': [True],
         'skip_double_backward_test': [True],
     })
 ))
-class TestIAdd(InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
+class TestIAdd(math_utils.InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a, b):
         a += b
@@ -529,7 +319,7 @@ def test_iadd_invalid_dtypes(device, dtypes):
         'skip_double_backward_test': [True],
     })
 ))
-class TestAddScalar(MathScalarTestBase, op_utils.NumpyOpTest):
+class TestAddScalar(math_utils.MathScalarTestBase, op_utils.NumpyOpTest):
 
     def func_scalar(self, xp, a, scalar):
         if self.is_module:
@@ -572,7 +362,8 @@ class TestAddScalar(MathScalarTestBase, op_utils.NumpyOpTest):
             0, -1, 1, 2, float('inf'), -float('inf'), float('nan')],
     })
 ))
-class TestIAddScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
+class TestIAddScalar(
+        math_utils.InplaceMathScalarTestBase, op_utils.NumpyOpTest):
 
     def func_scalar(self, xp, a, scalar):
         a += scalar
@@ -582,9 +373,10 @@ class TestIAddScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     # Special shapes
     chainer.testing.product({
-        'in_shapes': _shapes_combination_binary,
+        'in_shapes': math_utils.shapes_combination_binary,
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.numeric_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.numeric_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
         'is_module': [False],
@@ -601,7 +393,8 @@ class TestIAddScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.numeric_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.numeric_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
         'is_module': [True, False],
@@ -610,7 +403,8 @@ class TestIAddScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': ['random', float('inf'), -float('inf'), float('nan')],
         'input_rhs': ['random', float('inf'), -float('inf'), float('nan')],
         'is_module': [False],
@@ -618,7 +412,7 @@ class TestIAddScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
         'skip_double_backward_test': [True],
     })
 ))
-class TestSub(BinaryMathTestBase, op_utils.NumpyOpTest):
+class TestSub(math_utils.BinaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a, b):
         if self.is_module:
@@ -645,9 +439,10 @@ def test_sub_invalid_dtypes(device, dtypes, is_module):
 @chainer.testing.parameterize(*(
     # Special shapes
     chainer.testing.product({
-        'in_shapes': _shapes_combination_inplace_binary,
+        'in_shapes': math_utils.shapes_combination_inplace_binary,
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.numeric_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.numeric_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
     })
@@ -662,14 +457,15 @@ def test_sub_invalid_dtypes(device, dtypes, is_module):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': ['random', float('inf'), -float('inf'), float('nan')],
         'input_rhs': ['random', float('inf'), -float('inf'), float('nan')],
         'skip_backward_test': [True],
         'skip_double_backward_test': [True],
     })
 ))
-class TestISub(InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
+class TestISub(math_utils.InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a, b):
         a -= b
@@ -729,7 +525,7 @@ def test_isub_invalid_dtypes(device, dtypes):
         'skip_double_backward_test': [True],
     })
 ))
-class TestSubScalar(MathScalarTestBase, op_utils.NumpyOpTest):
+class TestSubScalar(math_utils.MathScalarTestBase, op_utils.NumpyOpTest):
 
     def func_scalar(self, xp, a, scalar):
         if self.is_module:
@@ -772,7 +568,8 @@ class TestSubScalar(MathScalarTestBase, op_utils.NumpyOpTest):
             0, -1, 1, 2, float('inf'), -float('inf'), float('nan')],
     })
 ))
-class TestISubScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
+class TestISubScalar(
+        math_utils.InplaceMathScalarTestBase, op_utils.NumpyOpTest):
 
     def func_scalar(self, xp, a, scalar):
         a -= scalar
@@ -782,9 +579,10 @@ class TestISubScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     # Special shapes
     chainer.testing.product({
-        'in_shapes': _shapes_combination_binary,
+        'in_shapes': math_utils.shapes_combination_binary,
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.all_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.all_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
         'is_module': [False],
@@ -801,7 +599,8 @@ class TestISubScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.all_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.all_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
         'is_module': [True, False],
@@ -810,7 +609,8 @@ class TestISubScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': ['random', float('inf'), -float('inf'), float('nan')],
         'input_rhs': ['random', float('inf'), -float('inf'), float('nan')],
         'is_module': [False],
@@ -818,7 +618,7 @@ class TestISubScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
         'skip_double_backward_test': [True],
     })
 ))
-class TestMul(BinaryMathTestBase, op_utils.NumpyOpTest):
+class TestMul(math_utils.BinaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a, b):
         if self.is_module:
@@ -831,9 +631,10 @@ class TestMul(BinaryMathTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     # Special shapes
     chainer.testing.product({
-        'in_shapes': _shapes_combination_inplace_binary,
+        'in_shapes': math_utils.shapes_combination_inplace_binary,
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.all_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.all_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
     })
@@ -850,14 +651,15 @@ class TestMul(BinaryMathTestBase, op_utils.NumpyOpTest):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': ['random', float('inf'), -float('inf'), float('nan')],
         'input_rhs': ['random', float('inf'), -float('inf'), float('nan')],
         'skip_backward_test': [True],
         'skip_double_backward_test': [True],
     })
 ))
-class TestIMul(InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
+class TestIMul(math_utils.InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a, b):
         a *= b
@@ -908,7 +710,7 @@ class TestIMul(InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
         'skip_double_backward_test': [True],
     })
 ))
-class TestMulScalar(MathScalarTestBase, op_utils.NumpyOpTest):
+class TestMulScalar(math_utils.MathScalarTestBase, op_utils.NumpyOpTest):
 
     def func_scalar(self, xp, a, scalar):
         if self.is_module:
@@ -953,7 +755,8 @@ class TestMulScalar(MathScalarTestBase, op_utils.NumpyOpTest):
             0, -1, 1, 2, float('inf'), -float('inf'), float('nan')],
     })
 ))
-class TestIMulScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
+class TestIMulScalar(
+        math_utils.InplaceMathScalarTestBase, op_utils.NumpyOpTest):
 
     def func_scalar(self, xp, a, scalar):
         a *= scalar
@@ -977,7 +780,7 @@ class TestIMulScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
     'in_dtypes,out_dtype': _in_out_dtypes_arithmetic,
     'is_module': [True, False],
 }))
-class TestFloorDivide(BinaryMathTestBase, op_utils.NumpyOpTest):
+class TestFloorDivide(math_utils.BinaryMathTestBase, op_utils.NumpyOpTest):
 
     skip_backward_test = True
     skip_double_backward_test = True
@@ -1020,7 +823,8 @@ class TestFloorDivide(BinaryMathTestBase, op_utils.NumpyOpTest):
             _in_out_dtypes_float_arithmetic_scalar,
     })
 )))
-class TestFloorDivideScalar(MathScalarTestBase, op_utils.NumpyOpTest):
+class TestFloorDivideScalar(
+        math_utils.MathScalarTestBase, op_utils.NumpyOpTest):
 
     skip_backward_test = True
     skip_double_backward_test = True
@@ -1164,7 +968,7 @@ _in_out_dtypes_truediv_scalar = _in_out_dtypes_inplace_truediv_scalar + [
 @chainer.testing.parameterize(*(
     # Special shapes
     chainer.testing.product({
-        'in_shapes': _shapes_combination_binary,
+        'in_shapes': math_utils.shapes_combination_binary,
         'in_dtypes,out_dtype': _in_out_dtypes_truediv,
         'input_lhs': ['random'],
         'input_rhs': ['random'],
@@ -1190,7 +994,8 @@ _in_out_dtypes_truediv_scalar = _in_out_dtypes_inplace_truediv_scalar + [
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': ['random', float('inf'), -float('inf'), float('nan')],
         'input_rhs': ['random', float('inf'), -float('inf'), float('nan')],
         'is_module': [False],
@@ -1198,7 +1003,7 @@ _in_out_dtypes_truediv_scalar = _in_out_dtypes_inplace_truediv_scalar + [
         'skip_double_backward_test': [True],
     })
 ))
-class TestTrueDivide(BinaryMathTestBase, op_utils.NumpyOpTest):
+class TestTrueDivide(math_utils.BinaryMathTestBase, op_utils.NumpyOpTest):
 
     check_numpy_strides_compliance = False
 
@@ -1215,7 +1020,7 @@ class TestTrueDivide(BinaryMathTestBase, op_utils.NumpyOpTest):
         a, b = super().generate_inputs()
         if self.input_lhs == 'random':
             # Avoid (-0.3, 0.3) interval
-            with IgnoreNumpyFloatingPointError():
+            with math_utils.IgnoreNumpyFloatingPointError():
                 b[numpy.logical_and(-0.3 < b, b < 0.3)] = 1
         return a, b
 
@@ -1244,7 +1049,7 @@ def test_truediv_invalid_dtypes(device, dtypes, is_module):
 @chainer.testing.parameterize(*(
     # Special shapes
     chainer.testing.product({
-        'in_shapes': _shapes_combination_inplace_binary,
+        'in_shapes': math_utils.shapes_combination_inplace_binary,
         'in_dtypes,out_dtype': _in_out_dtypes_inplace_truediv,
         'input_lhs': ['random'],
         'input_rhs': ['random'],
@@ -1260,14 +1065,16 @@ def test_truediv_invalid_dtypes(device, dtypes, is_module):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': ['random', float('inf'), -float('inf'), float('nan')],
         'input_rhs': ['random', float('inf'), -float('inf'), float('nan')],
         'skip_backward_test': [True],
         'skip_double_backward_test': [True],
     })
 ))
-class TestITrueDivide(InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
+class TestITrueDivide(
+        math_utils.InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
 
     skip_backward_test = True
     skip_double_backward_test = True
@@ -1275,7 +1082,7 @@ class TestITrueDivide(InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
     def generate_inputs(self):
         a, b = super().generate_inputs()
         if self.input_lhs == 'random':
-            with IgnoreNumpyFloatingPointError():
+            with math_utils.IgnoreNumpyFloatingPointError():
                 b[numpy.logical_and(-0.3 < b, b < 0.3)] = 1
         return a, b
 
@@ -1308,7 +1115,8 @@ class TestITrueDivide(InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
     + chainer.testing.product({
         'shape': [(2, 3)],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(1, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                1, chainerx.testing.float_dtypes)),
         'scalar_type': [float],
         'input': [float('inf'), -float('inf'), float('nan')],
         'scalar_value': [-1, 1, 2, float('inf'), -float('inf'), float('nan')],
@@ -1318,7 +1126,8 @@ class TestITrueDivide(InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
         'skip_double_backward_test': [True],
     })
 ))
-class TestTrueDivideScalar(MathScalarTestBase, op_utils.NumpyOpTest):
+class TestTrueDivideScalar(
+        math_utils.MathScalarTestBase, op_utils.NumpyOpTest):
 
     check_numpy_strides_compliance = False
 
@@ -1353,7 +1162,8 @@ class TestTrueDivideScalar(MathScalarTestBase, op_utils.NumpyOpTest):
     chainer.testing.product({
         'shape': [(), (0,), (1,), (2, 0, 3), (1, 1, 1), (2, 3)],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(1, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                1, chainerx.testing.float_dtypes)),
         'scalar_type': [float],
         'input': ['random'],
         'scalar_value': [1],
@@ -1362,363 +1172,18 @@ class TestTrueDivideScalar(MathScalarTestBase, op_utils.NumpyOpTest):
     + chainer.testing.product({
         'shape': [(2, 3)],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(1, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                1, chainerx.testing.float_dtypes)),
         'scalar_type': [float],
         'input': [float('inf'), -float('inf'), float('nan')],
         'scalar_value': [-1, 1, 2, float('inf'), -float('inf'), float('nan')],
     })
 ))
-class TestITrueDivideScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
+class TestITrueDivideScalar(
+        math_utils.InplaceMathScalarTestBase, op_utils.NumpyOpTest):
 
     def func_scalar(self, xp, a, scalar):
         a /= scalar
-
-
-_in_out_dtypes_bitwise = dtype_utils._permutate_dtype_mapping([
-    # Same dtypes
-    (('bool_', 'bool_'), 'bool_'),
-    (('int8', 'int8'), 'int8'),
-    (('int16', 'int16'), 'int16'),
-    (('int32', 'int32'), 'int32'),
-    (('int64', 'int64'), 'int64'),
-    (('uint8', 'uint8'), 'uint8'),
-    # Mixed dtypes
-    (('bool_', 'int8'), 'int8'),
-    (('bool_', 'int16'), 'int16'),
-    (('bool_', 'int32'), 'int32'),
-    (('bool_', 'int64'), 'int64'),
-    (('bool_', 'uint8'), 'uint8'),
-    (('int8', 'int16'), 'int16'),
-    (('int8', 'int32'), 'int32'),
-    (('int8', 'int64'), 'int64'),
-    (('int8', 'uint8'), 'int16'),
-    (('int16', 'int32'), 'int32'),
-    (('int16', 'int64'), 'int64'),
-    (('int16', 'uint8'), 'int16'),
-    (('int32', 'int64'), 'int64'),
-    (('int32', 'uint8'), 'int32'),
-    (('int64', 'uint8'), 'int64'),
-])
-
-_inplace_invalid_bitwise = [
-    (('bool_', 'int8'), 'int8'),
-    (('bool_', 'int16'), 'int16'),
-    (('bool_', 'int32'), 'int32'),
-    (('bool_', 'int64'), 'int64'),
-    (('bool_', 'uint8'), 'uint8'),
-]
-
-_in_out_inplace_dtypes_bitwise = [
-    dtypes for dtypes in _in_out_dtypes_bitwise
-    if dtypes not in _inplace_invalid_bitwise
-]
-
-_scalar_invalid_bitwise = [
-    (('float16',), int, 'float16'),
-    (('float32',), int, 'float32'),
-    (('float64',), int, 'float64'),
-    (('float64',), numpy.int8, 'float64'),
-    (('float16',), numpy.int64, 'float16'),
-]
-
-_in_out_scalar_dtypes_bitwise = [
-    dtypes for dtypes in _in_out_dtypes_array_int_scalar
-    if dtypes not in _scalar_invalid_bitwise
-]
-
-_params_bitwise = (
-    # Special shapes
-    chainer.testing.product({
-        'in_shapes': _shapes_combination_binary,
-        'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.nonfloat_dtypes)),
-        'input_lhs': ['random'],
-        'input_rhs': ['random'],
-        'is_module': [False],
-    })
-    # Dtype combinations
-    + chainer.testing.product({
-        'in_shapes': [((2, 3), (2, 3))],
-        'in_dtypes,out_dtype': _in_out_dtypes_bitwise,
-        'input_lhs': ['random'],
-        'input_rhs': ['random'],
-        'is_module': [False],
-    })
-    # is_module
-    + chainer.testing.product({
-        'in_shapes': [((2, 3), (2, 3))],
-        'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.nonfloat_dtypes)),
-        'input_lhs': ['random'],
-        'input_rhs': ['random'],
-        'is_module': [True, False],
-    })
-)
-
-
-_inplace_params_bitwise = (
-    # Special shapes
-    chainer.testing.product({
-        'in_shapes': _shapes_combination_inplace_binary,
-        'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.nonfloat_dtypes)),
-        'input_lhs': ['random'],
-        'input_rhs': ['random'],
-    })
-    # Dtype combinations
-    + chainer.testing.product({
-        'in_shapes': [((2, 3), (2, 3))],
-        'in_dtypes,out_dtype': _in_out_inplace_dtypes_bitwise,
-        'input_lhs': ['random'],
-        'input_rhs': ['random'],
-    })
-    # Special values
-    + chainer.testing.product({
-        'in_shapes': [((2, 3), (2, 3))],
-        'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.nonfloat_dtypes)),
-        'input_lhs': ['random', float('inf'), -float('inf'), float('nan')],
-        'input_rhs': ['random', float('inf'), -float('inf'), float('nan')],
-    })
-)
-
-
-_scalar_params_bitwise = (
-    # Special shapes
-    chainer.testing.product({
-        'shape': [(), (0,), (1,), (2, 0, 3), (1, 1, 1), (2, 3)],
-        'in_dtypes,scalar_type,out_dtype': _in_out_scalar_dtypes_bitwise,
-        'input': ['random'],
-        'scalar_value': [1],
-        'is_module': [False],
-        'is_scalar_rhs': [False],
-    })
-    # Type combinations
-    + chainer.testing.product({
-        'shape': [(2, 3)],
-        'in_dtypes,scalar_type,out_dtype': _in_out_scalar_dtypes_bitwise,
-        'input': ['random'],
-        'scalar_value': [1],
-        'is_module': [False],
-        'is_scalar_rhs': [True, False],
-    })
-    # is_module
-    + chainer.testing.product({
-        'shape': [(2, 3)],
-        'in_dtypes,scalar_type,out_dtype': _in_out_scalar_dtypes_bitwise,
-        'input': ['random'],
-        'scalar_value': [1],
-        'is_module': [True, False],
-        'is_scalar_rhs': [True, False],
-    })
-    # Special values
-    + chainer.testing.product({
-        'shape': [(2, 3)],
-        'in_dtypes,scalar_type,out_dtype':
-            _in_out_scalar_dtypes_bitwise,
-        'input': [float('inf'), -float('inf'), float('nan')],
-        'scalar_value': [
-            0, -1, 1, 2],
-        'is_module': [False],
-        'is_scalar_rhs': [False],
-        'skip_backward_test': [True],
-        'skip_double_backward_test': [True],
-    })
-)
-
-
-_inplace_scalar_params_bitwise = (
-    # Special shapes
-    chainer.testing.product({
-        'shape': [(), (0,), (1,), (2, 0, 3), (1, 1, 1), (2, 3)],
-        'in_dtypes,scalar_type,out_dtype':
-            _in_out_scalar_dtypes_bitwise,
-        'input': ['random'],
-        'scalar_value': [1],
-    })
-    # Dtype combinations
-    + chainer.testing.product({
-        'shape': [(2, 3)],
-        'in_dtypes,scalar_type,out_dtype':
-            _in_out_scalar_dtypes_bitwise,
-        'input': ['random'],
-        'scalar_value': [1],
-    })
-    # Special values
-    + chainer.testing.product({
-        'shape': [(2, 3)],
-        'in_dtypes,scalar_type,out_dtype':
-            _in_out_scalar_dtypes_bitwise,
-        'input': [float('inf'), -float('inf'), float('nan')],
-        'scalar_value': [
-            0, -1, 1, 2],
-    })
-)
-
-
-@op_utils.op_test(['native:0', 'cuda:0'])
-@chainer.testing.parameterize(*_params_bitwise)
-class TestBitwiseAnd(BinaryMathTestBase, op_utils.NumpyOpTest):
-
-    def func(self, xp, a, b):
-        if self.is_module:
-            return xp.bitwise_and(a, b)
-        else:
-            return a & b
-
-
-@op_utils.op_test(['native:0', 'cuda:0'])
-@chainer.testing.parameterize(*_params_bitwise)
-class TestBitwiseOr(BinaryMathTestBase, op_utils.NumpyOpTest):
-
-    def func(self, xp, a, b):
-        if self.is_module:
-            return xp.bitwise_or(a, b)
-        else:
-            return a | b
-
-
-@op_utils.op_test(['native:0', 'cuda:0'])
-@chainer.testing.parameterize(*_params_bitwise)
-class TestBitwiseXor(BinaryMathTestBase, op_utils.NumpyOpTest):
-
-    def func(self, xp, a, b):
-        if self.is_module:
-            return xp.bitwise_xor(a, b)
-        else:
-            return a ^ b
-
-
-@op_utils.op_test(['native:0', 'cuda:0'])
-@chainer.testing.parameterize(*_inplace_params_bitwise)
-class TestIBitwiseAnd(InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
-
-    def func(self, xp, a, b):
-        a &= b
-
-
-@op_utils.op_test(['native:0', 'cuda:0'])
-@chainer.testing.parameterize(*_inplace_params_bitwise)
-class TestIBitwiseOr(InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
-
-    def func(self, xp, a, b):
-        a |= b
-
-
-@op_utils.op_test(['native:0', 'cuda:0'])
-@chainer.testing.parameterize(*_inplace_params_bitwise)
-class TestIBitwiseXor(InplaceBinaryMathTestBase, op_utils.NumpyOpTest):
-
-    def func(self, xp, a, b):
-        a ^= b
-
-
-@pytest.mark.parametrize_device(['native:0', 'cuda:0'])
-@pytest.mark.parametrize('dtypes', _inplace_invalid_bitwise)
-def test_iand_invalid_dtypes(device, dtypes):
-    (in_dtype1, in_dtype2), _ = dtypes
-    shape = (2, 3)
-    a = chainerx.array(array_utils.uniform(shape, in_dtype1))
-    b = chainerx.array(array_utils.uniform(shape, in_dtype2))
-    with pytest.raises(chainerx.DtypeError):
-        a &= b
-
-
-@pytest.mark.parametrize_device(['native:0', 'cuda:0'])
-@pytest.mark.parametrize('dtypes', _inplace_invalid_bitwise)
-def test_ior_invalid_dtypes(device, dtypes):
-    (in_dtype1, in_dtype2), _ = dtypes
-    shape = (2, 3)
-    a = chainerx.array(array_utils.uniform(shape, in_dtype1))
-    b = chainerx.array(array_utils.uniform(shape, in_dtype2))
-    with pytest.raises(chainerx.DtypeError):
-        a |= b
-
-
-@pytest.mark.parametrize_device(['native:0', 'cuda:0'])
-@pytest.mark.parametrize('dtypes', _inplace_invalid_bitwise)
-def test_ixor_invalid_dtypes(device, dtypes):
-    (in_dtype1, in_dtype2), _ = dtypes
-    shape = (2, 3)
-    a = chainerx.array(array_utils.uniform(shape, in_dtype1))
-    b = chainerx.array(array_utils.uniform(shape, in_dtype2))
-    with pytest.raises(chainerx.DtypeError):
-        a ^= b
-
-
-@op_utils.op_test(['native:0', 'cuda:0'])
-@chainer.testing.parameterize(*_scalar_params_bitwise)
-class TestBitwiseAndScalar(MathScalarTestBase, op_utils.NumpyOpTest):
-
-    def func_scalar(self, xp, a, scalar):
-        if self.is_module:
-            if self.is_scalar_rhs:
-                return a & scalar
-            else:
-                return scalar & a
-        else:
-            if self.is_scalar_rhs:
-                return xp.bitwise_and(a, scalar)
-            else:
-                return xp.bitwise_and(scalar, a)
-
-
-@op_utils.op_test(['native:0', 'cuda:0'])
-@chainer.testing.parameterize(*_scalar_params_bitwise)
-class TestBitwiseOrScalar(MathScalarTestBase, op_utils.NumpyOpTest):
-
-    def func_scalar(self, xp, a, scalar):
-        if self.is_module:
-            if self.is_scalar_rhs:
-                return a | scalar
-            else:
-                return scalar | a
-        else:
-            if self.is_scalar_rhs:
-                return xp.bitwise_or(a, scalar)
-            else:
-                return xp.bitwise_or(scalar, a)
-
-
-@op_utils.op_test(['native:0', 'cuda:0'])
-@chainer.testing.parameterize(*_scalar_params_bitwise)
-class TestBitwiseXorScalar(MathScalarTestBase, op_utils.NumpyOpTest):
-
-    def func_scalar(self, xp, a, scalar):
-        if self.is_module:
-            if self.is_scalar_rhs:
-                return a ^ scalar
-            else:
-                return scalar ^ a
-        else:
-            if self.is_scalar_rhs:
-                return xp.bitwise_xor(a, scalar)
-            else:
-                return xp.bitwise_xor(scalar, a)
-
-
-@op_utils.op_test(['native:0', 'cuda:0'])
-@chainer.testing.parameterize(*_inplace_scalar_params_bitwise)
-class TestIBitwiseAndScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
-
-    def func_scalar(self, xp, a, scalar):
-        a &= scalar
-
-
-@op_utils.op_test(['native:0', 'cuda:0'])
-@chainer.testing.parameterize(*_inplace_scalar_params_bitwise)
-class TestIBitwiseOrScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
-
-    def func_scalar(self, xp, a, scalar):
-        a |= scalar
-
-
-@op_utils.op_test(['native:0', 'cuda:0'])
-@chainer.testing.parameterize(*_inplace_scalar_params_bitwise)
-class TestIBitwiseXorScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
-
-    def func_scalar(self, xp, a, scalar):
-        a ^= scalar
 
 
 @op_utils.op_test(['native:0', 'cuda:0'])
@@ -1766,7 +1231,7 @@ class TestIBitwiseXorScalar(InplaceMathScalarTestBase, op_utils.NumpyOpTest):
 ])
 @chainer.testing.parameterize_pytest('keepdims', [True, False])
 @chainer.testing.parameterize_pytest('is_module', [True, False])
-class TestSum(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestSum(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     input = 'random'
 
@@ -1881,7 +1346,7 @@ def test_sum_invalid(is_module, xp, shape, axis, keepdims, dtype):
         'skip_double_backward_test': [True],
     })
 ))
-class TestMinimumScalar(MathScalarTestBase, op_utils.NumpyOpTest):
+class TestMinimumScalar(math_utils.MathScalarTestBase, op_utils.NumpyOpTest):
 
     dodge_nondifferentiable = True
 
@@ -1930,7 +1395,7 @@ class TestMinimumScalar(MathScalarTestBase, op_utils.NumpyOpTest):
         'skip_double_backward_test': [True],
     })
 ))
-class TestMaximumScalar(MathScalarTestBase, op_utils.NumpyOpTest):
+class TestMaximumScalar(math_utils.MathScalarTestBase, op_utils.NumpyOpTest):
 
     dodge_nondifferentiable = True
 
@@ -2032,7 +1497,7 @@ _in_out_dtypes_math_binary_functions = dtype_utils._permutate_dtype_mapping([
         'skip_double_backward_test': [True],
     })
 ))
-class TestExp(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestExp(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.exp(a)
@@ -2062,7 +1527,7 @@ class TestExp(UnaryMathTestBase, op_utils.NumpyOpTest):
         'skip_double_backward_test': [True],
     })
 ))
-class TestLog(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestLog(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.log(a)
@@ -2092,7 +1557,7 @@ class TestLog(UnaryMathTestBase, op_utils.NumpyOpTest):
         'skip_double_backward_test': [True],
     })
 ))
-class TestLog10(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestLog10(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.log10(a)
@@ -2131,7 +1596,7 @@ _invalid_logsumexp_params = [
     'in_dtypes,out_dtype', _in_out_dtypes_math_functions)
 @chainer.testing.parameterize_pytest('shape,axis', _logsumexp_params)
 @chainer.testing.parameterize_pytest('keepdims', [True, False])
-class TestLogSumExp(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestLogSumExp(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     input = 'random'
 
@@ -2165,7 +1630,7 @@ def test_logsumexp_invalid(device, a_shape, axis, keepdims, dtype):
 @chainer.testing.parameterize_pytest('shape,axis', _logsumexp_params)
 @chainer.testing.parameterize_pytest(
     'in_dtypes,out_dtype', _in_out_dtypes_math_functions)
-class TestLogSoftmax(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestLogSoftmax(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     input = 'random'
 
@@ -2198,7 +1663,8 @@ def test_log_softmax_invalid(device, a_shape, axis, dtype):
     chainer.testing.product({
         'shape': [(), (0,), (1,), (2, 0, 3), (1, 1, 1), (2, 3)],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
     })
@@ -2206,7 +1672,8 @@ def test_log_softmax_invalid(device, a_shape, axis, dtype):
     + chainer.testing.product({
         'shape': [(2, 3)],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': ['random', float('inf'), -float('inf'), float('nan')],
         'input_rhs': ['random', float('inf'), -float('inf'), float('nan')],
         'skip_backward_test': [True],
@@ -2292,7 +1759,7 @@ class TestSigmoid(op_utils.NumpyOpTest):
 @chainer.testing.parameterize_pytest('shape,axis', _logsumexp_params)
 @chainer.testing.parameterize_pytest(
     'in_dtypes,out_dtype', _in_out_dtypes_math_functions)
-class TestSoftmax(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestSoftmax(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     input = 'random'
 
@@ -2329,7 +1796,7 @@ class TestSoftmax(UnaryMathTestBase, op_utils.NumpyOpTest):
         'skip_double_backward_test': [True],
     })
 ))
-class TestSquare(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestSquare(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.square(a)
@@ -2359,7 +1826,7 @@ class TestSquare(UnaryMathTestBase, op_utils.NumpyOpTest):
         'skip_double_backward_test': [True],
     })
 ))
-class TestSqrt(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestSqrt(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.sqrt(a)
@@ -2384,7 +1851,7 @@ _trigonometric_hyperbolic_params = \
 @chainer.testing.parameterize(*(
     _trigonometric_hyperbolic_params
 ))
-class TestSinh(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestSinh(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.sinh(a)
@@ -2394,7 +1861,7 @@ class TestSinh(UnaryMathTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     _trigonometric_hyperbolic_params
 ))
-class TestCosh(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestCosh(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.cosh(a)
@@ -2404,7 +1871,7 @@ class TestCosh(UnaryMathTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     _trigonometric_hyperbolic_params
 ))
-class TestTanh(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestTanh(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.tanh(a)
@@ -2414,7 +1881,7 @@ class TestTanh(UnaryMathTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     _trigonometric_hyperbolic_params
 ))
-class TestSin(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestSin(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.sin(a)
@@ -2424,7 +1891,7 @@ class TestSin(UnaryMathTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     _trigonometric_hyperbolic_params
 ))
-class TestCos(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestCos(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.cos(a)
@@ -2434,7 +1901,7 @@ class TestCos(UnaryMathTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     _trigonometric_hyperbolic_params
 ))
-class TestTan(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestTan(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     dodge_nondifferentiable = True
     check_backward_options = {'atol': 3e-5}
@@ -2459,7 +1926,7 @@ class TestTan(UnaryMathTestBase, op_utils.NumpyOpTest):
         'skip_double_backward_test': [True],
     })
 ))
-class TestAbs(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestAbs(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     dodge_nondifferentiable = True
 
@@ -2516,7 +1983,7 @@ def _make_inverse_trig_params(name):
 @chainer.testing.parameterize(*(
     _make_inverse_trig_params('arcsinh')
 ))
-class TestArcsinh(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestArcsinh(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.arcsinh(a)
@@ -2526,7 +1993,7 @@ class TestArcsinh(UnaryMathTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     _make_inverse_trig_params('arccosh')
 ))
-class TestArccosh(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestArccosh(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.arccosh(a)
@@ -2536,7 +2003,7 @@ class TestArccosh(UnaryMathTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     _make_inverse_trig_params('arcsin')
 ))
-class TestArcsin(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestArcsin(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.arcsin(a)
@@ -2546,7 +2013,7 @@ class TestArcsin(UnaryMathTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     _make_inverse_trig_params('arccos')
 ))
-class TestArccos(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestArccos(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.arccos(a)
@@ -2556,7 +2023,7 @@ class TestArccos(UnaryMathTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     _make_inverse_trig_params('arctan')
 ))
-class TestArctan(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestArctan(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.arctan(a)
@@ -2568,9 +2035,10 @@ class TestArctan(UnaryMathTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     # Special shapes
     chainer.testing.product({
-        'in_shapes': _shapes_combination_binary,
+        'in_shapes': math_utils.shapes_combination_binary,
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': [1],
         'input_rhs': [2],
         'skip_backward_test': [True],
@@ -2580,7 +2048,8 @@ class TestArctan(UnaryMathTestBase, op_utils.NumpyOpTest):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': [-3, -0.75, 0.75, 3],
         'input_rhs': [-3, -0.75, 0.75, 3],
     })
@@ -2595,7 +2064,8 @@ class TestArctan(UnaryMathTestBase, op_utils.NumpyOpTest):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.float_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.float_dtypes)),
         'input_lhs': ['random', float('inf'), -float('inf'), float('nan'),
                       +0.0, -0.0],
         'input_rhs': ['random', float('inf'), -float('inf'), float('nan'),
@@ -2604,7 +2074,7 @@ class TestArctan(UnaryMathTestBase, op_utils.NumpyOpTest):
         'skip_double_backward_test': [True],
     })
 ))
-class TestArctan2(BinaryMathTestBase, op_utils.NumpyOpTest):
+class TestArctan2(math_utils.BinaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a, b):
         return xp.arctan2(a, b)
@@ -2628,7 +2098,7 @@ class TestArctan2(BinaryMathTestBase, op_utils.NumpyOpTest):
         'skip_double_backward_test': [True],
     })
 ))
-class TestFabs(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestFabs(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.fabs(a)
@@ -2777,19 +2247,21 @@ _minmax_params = [
             ((4, 3), (-2, -1)),
         ],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(1, chainerx.testing.all_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                1, chainerx.testing.all_dtypes)),
         'is_module': [True, False],
     }) +
     chainer.testing.product({
         'array,axis': _minmax_params,
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(1, chainerx.testing.all_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                1, chainerx.testing.all_dtypes)),
         'is_module': [True, False],
         'skip_backward_test': [True],
         'skip_double_backward_test': [True],
     })
 ))
-class TestMax(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestMax(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     dodge_nondifferentiable = True
 
@@ -2844,19 +2316,21 @@ def test_min_amin():
             ((4, 3), (-2, -1)),
         ],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(1, chainerx.testing.all_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                1, chainerx.testing.all_dtypes)),
         'is_module': [True, False],
     }) +
     chainer.testing.product({
         'array,axis': _minmax_params,
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(1, chainerx.testing.all_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                1, chainerx.testing.all_dtypes)),
         'is_module': [True, False],
         'skip_backward_test': [True],
         'skip_double_backward_test': [True],
     })
 ))
-class TestMin(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestMin(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     dodge_nondifferentiable = True
 
@@ -2896,9 +2370,10 @@ def test_min_invalid_shapes_and_axis(device, array, axis, dtype, is_module):
 @chainer.testing.parameterize(*(
     # Special shapes
     chainer.testing.product({
-        'in_shapes': _shapes_combination_binary,
+        'in_shapes': math_utils.shapes_combination_binary,
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.numeric_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.numeric_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
     })
@@ -2914,14 +2389,15 @@ def test_min_invalid_shapes_and_axis(device, array, axis, dtype, is_module):
     + chainer.testing.product({
         'in_shapes': [((2, 3), (2, 3))],
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.numeric_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.numeric_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
         'is_module': [True, False],
     })
     # TODO(aksub99): Add tests for inf and NaN.
 ))
-class TestMaximum(BinaryMathTestBase, op_utils.NumpyOpTest):
+class TestMaximum(math_utils.BinaryMathTestBase, op_utils.NumpyOpTest):
 
     dodge_nondifferentiable = True
 
@@ -2944,9 +2420,10 @@ def test_maximum_invalid_dtypes(device, dtypes):
 @chainer.testing.parameterize(*(
     # Special shapes
     chainer.testing.product({
-        'in_shapes': _shapes_combination_binary,
+        'in_shapes': math_utils.shapes_combination_binary,
         'in_dtypes,out_dtype': (
-            _make_same_in_out_dtypes(2, chainerx.testing.numeric_dtypes)),
+            dtype_utils.make_same_in_out_dtypes(
+                2, chainerx.testing.numeric_dtypes)),
         'input_lhs': ['random'],
         'input_rhs': ['random'],
         'is_module': [False],
@@ -2960,7 +2437,7 @@ def test_maximum_invalid_dtypes(device, dtypes):
     })
     # TODO(aksub99): Add tests for inf and NaN.
 ))
-class TestMinimum(BinaryMathTestBase, op_utils.NumpyOpTest):
+class TestMinimum(math_utils.BinaryMathTestBase, op_utils.NumpyOpTest):
 
     dodge_nondifferentiable = True
 
@@ -3006,7 +2483,7 @@ _mean_var_params = \
 @chainer.testing.parameterize(*(
     _mean_var_params
 ))
-class TestMean(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestMean(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.mean(a, self.axis)
@@ -3016,7 +2493,7 @@ class TestMean(UnaryMathTestBase, op_utils.NumpyOpTest):
 @chainer.testing.parameterize(*(
     _mean_var_params
 ))
-class TestVar(UnaryMathTestBase, op_utils.NumpyOpTest):
+class TestVar(math_utils.UnaryMathTestBase, op_utils.NumpyOpTest):
 
     def func(self, xp, a):
         return xp.var(a, self.axis)
